ints and chars

+++ src/Language/PureScript/Interpreter/JSON.hs
@@ -19,7 +19,7 @@ fromJSON :: Aeson.Value -> Evaluate.Value m
 fromJSON (Aeson.Object x) = Evaluate.Object (fmap fromJSON x)
 fromJSON (Aeson.Array x) = Evaluate.Array (fmap fromJSON x)
 fromJSON (Aeson.String x) = Evaluate.String x
-fromJSON (Aeson.Number x) = Evaluate.Number x
+fromJSON (Aeson.Number x) = Evaluate.Number (realToFrac x)
 fromJSON (Aeson.Bool x) = Evaluate.Bool x
 fromJSON Aeson.Null = error "TODO: not supported"
 
@@ -27,7 +27,7 @@ toJSON :: Evaluate.Value m -> Maybe Aeson.Value
 toJSON (Evaluate.Object x) = Aeson.Object <$> traverse toJSON x
 toJSON (Evaluate.Array x) = Aeson.Array <$> traverse toJSON x
 toJSON (Evaluate.String x) = pure (Aeson.String x)
-toJSON (Evaluate.Number x) = pure (Aeson.Number x)
+toJSON (Evaluate.Number x) = pure (Aeson.Number (realToFrac x))
 toJSON (Evaluate.Bool x) = pure (Aeson.Bool x)
 toJSON _ = Nothing
